Separate server handlers into separate files, add middleware, and clean up server files
