+++ pkg/server/handlers/auth.go
@@ -1,4 +1,4 @@
-package server
+package handlers
 
 import (
 	"database/sql"
@@ -11,7 +11,12 @@ import (
 	"github.com/ellismcdougald/edmonton-bike-map/pkg/utils"
 )
 
-func handleLogin(writer http.ResponseWriter, request *http.Request, db *sql.DB) {
+// HandleLogin handles HTTP requests to log in to the application.
+// The request body includes the given username and password
+// The username and password is validated against existing usernames and (hashed) passwords
+// Responds with a generated JWT token if login is successful
+// Returns HTTP 401 if username or password is rejected
+func HandleLogin(writer http.ResponseWriter, request *http.Request, db *sql.DB) {
 	if request.Method != http.MethodPost {
 		http.Error(writer, "Only POST allowed", http.StatusMethodNotAllowed)
 		return
@@ -42,12 +47,20 @@ func handleLogin(writer http.ResponseWriter, request *http.Request, db *sql.DB) 
 	}
 
 	writer.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(writer).Encode(map[string]string{
+	err = json.NewEncoder(writer).Encode(map[string]string{
 		"token": token,
 	})
+	if err != nil {
+		http.Error(writer, "Could not encode JWT token", http.StatusInternalServerError)
+		return
+	}
 }
 
-func handleSignUp(writer http.ResponseWriter, request *http.Request, db *sql.DB) {
+// HandleSignUp handles HTTP requests to sign up for the application
+// Request body includes proposed username and password
+// If username does not already exist, then the new user is created
+// Returns HTTP 201 on success.
+func HandleSignUp(writer http.ResponseWriter, request *http.Request, db *sql.DB) {
 	if request.Method != http.MethodPost {
 		http.Error(writer, "Only POST allowed", http.StatusMethodNotAllowed)
 		return
